Attemtpt to improve decoder
I noticed that the 4 bits lead in vs 3 bits lead in isn;t enough because
in `ts_GetCommnand` pulses 5 times then reads 8 bits to read a command.

The comments call this a RD operation but I don't understand how this
fits into the 4 bit lead in also labelled RD earlier so we add a
seperate case.