Add comment about a different way to decode
I wonder if the protocol is just 1-4 pulses:

1 pulse outputs a `1` data bit
2 pulses outputs a `0` data bit
3 pulses starts a `WS` operation
4 pulses starts a `RD` operation

When there is a `RD` operation the the adpater keeps sending bits as
long as the host keeps sending pulses.

When there is a `WS` operation then the host sends bits.

When a large delay has passed (the inter bit delay) then the state
resets to looking for a new operation.

This is fairly neat and simple, it could be the intention.