Update notes with information from service manual
This gives us expect timings so we can make the decoder more robust and
allow it to skip the ROM pattern activity that would otherwise confuse
the somewhat naive implement of the decoder I wrote.

Using this timing information we should be able to make the decoder much
more robust.