Refactor acorn_post
I liked the pattern of filling a buffer and letting decoding functions
clear the buffer if they find valid inputs.

We can clean up the pulse detection code because this pattern gives us
a way to implement look ahead functionality, if the decoder needs to
look ahead but not consume a pulse it can leave it in the buffer for
subseequence decoders to read.

This makes the code less convoluted for the case of outputs that are
terminated with a input.